requirements.txt
aiogram==2.19
aiohttp==3.8.1
aioredis==2.0.1
aiosignal==1.2.0
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
bidict==0.22.0
certifi==2021.10.8
charset-normalizer==2.0.12
Deprecated==1.2.13
Django==4.0.3
django-dramatiq==0.10.0
django-ninja==0.17.0
django-resized==0.3.11
dramatiq==1.12.3
frozenlist==1.3.0
idna==3.3
jmespath==1.0.0
multidict==6.0.2
packaging==21.3
pika==1.2.0
Pillow==9.0.1
prometheus-client==0.13.1
pydantic==1.9.0
pyparsing==3.0.8
python-dateutil==2.8.2
python-engineio==4.3.2
python-slugify==6.1.1
python-socketio==5.6.0
pytz==2022.1
redis==4.2.2
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
typing_extensions==4.1.1
tzdata==2022.1
wrapt==1.14.0
yarl==1.7.2
